ai-geomarket-advisor-agent/geomarket-advisor-runtime/requirements.txt
google-adk
google-generativeai
mcp
fastapi
uvicorn
python-dotenv
httpx
pydantic

